+++ src/main/java/dev/coder2195/stellarity/interface_injection/ExtEntity.java
@@ -5,8 +5,18 @@ import dev.coder2195.stellarity.registry.StellarityDataAttachments;
 
 @SuppressWarnings("NonExtendableApiUsage")
 public interface ExtEntity extends AttachmentTarget {
+	void setGlowingTag(boolean glowing);
+
+	default void stellarity$removeGlowColor() {
+		this.removeAttached(StellarityDataAttachments.GLOW_COLOR);
+	}
 
 	default void stellarity$setGlowColor(int color) {
+		setGlowingTag(color != -1);
+		if (color == -1) {
+			stellarity$removeGlowColor();
+			return;
+		}
 		this.setAttached(StellarityDataAttachments.GLOW_COLOR, color);
 	}